requirements.txt
openai==1.98.0
prettyprint==0.1.5
python-dotenv==1.1.1
textworld==1.6.2